# lab1/docs/swagger.yaml
basePath: /api
definitions:
  api.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  api.LoginResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/ds.MedUserResponse'
    type: object
  api.LogoutRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  ds.MedUserResponse:
    properties:
      id:
        type: integer
      is_moderator:
        type: boolean
      login:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@lungcalc.ru
    name: API Support
    url: http://localhost:8080
  description: |-
    API для расчета должной жизненной емкости легких (ДЖЕЛ)
    Лабораторная работа 4 - Добавление аутентификации и авторизации
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Lung Capacity Calculation API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Выполняет вход пользователя и возвращает JWT токен
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Аутентификация пользователя
      tags:
      - auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Добавляет JWT токен в черный список
      parameters:
      - description: Токен для выхода
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Выход пользователя
      tags:
      - auth
  /api/auth/profile:
    get:
      description: Возвращает информацию о текущем пользователе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.MedUserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение профиля пользователя
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен для аутентификации. Формат: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"